% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evaluate}
\alias{evaluate}
\title{Do evaluation.}
\usage{
evaluate(pop, model_data, model = "lm", criterion = "AIC",
  do_parallel = FALSE)
}
\arguments{
\item{pop}{Matrix of population of chromosomes.}

\item{do_parallel}{Logical; Default \code{FALSE}; Do in parallel?}

\item{model_data;}{Object of class \code{model_data}.}

\item{model;}{Character; "lm" (default) or "glm"; Linear model or
generalized linear model.}

\item{criterion;}{"AIC" (default) or "BIC"; Criterion to be minimized.}
}
\value{
Numeric vector; Evaluation values for all chromosomes
in the current generation.
}
\description{
Do evaluation for chromosomes in population by calculating model selection criterion.
}

