% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evaluate_once}
\alias{evaluate_once}
\title{Do evaluation once.}
\usage{
evaluate_once(model_data, xvars_select, model = "lm", criterion = "AIC")
}
\arguments{
\item{model_data;}{Object of class \code{model_data}.}

\item{xvars_select;}{Logical vector;}

\item{model;}{Character; "lm" (default) or "glm"; Linear model or generalized linear model.}

\item{criterion;}{"AIC" (default) or "BIC"; AIC or BIC.}
}
\value{
Numeric; Value of criterion.
}
\description{
Do evaluation for a chromosome by calculating model selection criterion.
}

