% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mutate}
\alias{mutate}
\title{Mutate genes in the population.}
\usage{
mutate(pop, prob_mutate = 0.01)
}
\arguments{
\item{pop}{Matrix; Population of chromosomes.}

\item{prob_mutate}{Numeric, between 0 and 1; Default is 0.01; Probability of mutation.}
}
\value{
Matrix of population of chromosomes that have undergone mutation.
}
\description{
Mutate each gene in the population at a pre-defined rate.
}

