% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recombine_once}
\alias{recombine_once}
\title{Recombine once.}
\usage{
recombine_once(parent1, parent2, method = "onepoint")
}
\arguments{
\item{parent1}{Integer vector of 1st parent chromosome containing 1's and 0's.}

\item{parent2}{Integer vector of 2nd parent chromosome containing 1's and 0's.}

\item{method}{String; "onepoint" (default), "twopoint", "uniform";
Type of crossover, at one point, at two points or uniformly (at all possible points).}
}
\value{
Integer vector of child chromosome containing 1's and 0's.
}
\description{
Carry out crossover of two parent chromosomes to produce one child chromosome.
}

