% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select}
\alias{select}
\title{Select chromosomes for recombination.}
\usage{
select(pop, evaluation, method = "rank", do_parallel = FALSE)
}
\arguments{
\item{pop}{Matrix; Population of chromosomes.}

\item{evaluation}{Numeric vector; Evaluation values of all chromosomes in population.}

\item{method}{String; "rank" (linear rank selection) (default) or
"tournament"; Method to select chromosomes for inclusion in mating pool.}

\item{do_parallel}{Logical; Default \code{FALSE}; Do in parallel?}
}
\value{
Matrix of population of chromosomes that form the mating pool.
}
\description{
Select chromosomes for recombination based on fitness.
}

